// libs/common/ui-kit/src/lib/feature/navigation/navigation.interfaces.ts
export interface SeekNavigationItem {
  id: string;
  title: string;
  type: 'item' | 'group' | 'collapsable';
  translate?: string;
  icon?: string;
  hidden?: boolean;
  url?: string;
  classes?: string;
  exactMatch?: boolean;
  externalUrl?: boolean;
  openInNewTab?: boolean;
  function?: any;
  badge?: {
    title?: string;
    translate?: string;
    bg?: string;
    fg?: string;
  };
  children?: SeekNavigationItem[];
}

export interface SeekNavigation extends SeekNavigationItem {
  children?: SeekNavigationItem[];
}